backend/groq-api/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
python-dotenv==1.0.1
groq==0.4.2
python-multipart==0.0.9
pydantic==2.6.3
loguru==0.7.2
httpx==0.27.0
ffmpeg-python==0.2.0 